Split up PIE tests.
It looks like on ARM-64, all binaries are position-independent, so this test
only succeeds on x86-64.
