+++ tests/core/go_binary/BUILD.bazel
@@ -95,6 +95,7 @@ go_test(
     name = "pie_test",
     srcs = [
         "pie_darwin_test.go",
+        "pie_darwin_amd64_test.go",
         "pie_linux_test.go",
     ],
     data = select({
