+++ tests/core/go_binary/pie_darwin_amd64_test.go
@@ -1,0 +1,31 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package test
+
+import (
+	"debug/macho"
+	"testing"
+)
+
+func TestNoPIE(t *testing.T) {
+	m, err := openMachO("tests/core/go_binary", "hello_nopie_bin")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if m.Flags&macho.FlagPIE != 0 {
+		t.Error("ELF binary is not position-dependent.")
+	}
+}
